==> kaggle/08 - Competitions/02 - Jigsaw toxic severity rating/08_dense_vectors.py <==
# Check this kaggle notebook which use transformers from
# hugging face and convert my sentences in a matrix of
# vectors with 768 features. I use kaggle because my powerfull 
# intel pentium inside said no more!
# https://www.kaggle.com/estebanmarcelloni/jigsaw-toxic-rating-using-setence2vector
